# highlight-erp/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Highlight ERP API Documentation",
        "description": "Документация API для проекта Highlight ERP",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
        "/api/admin/documents": {
            "get": {
                "tags": [
                    "Администратор - Документы"
                ],
                "summary": "Получение списка всех документов",
                "operationId": "getDocuments",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Document"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Администратор - Документы"
                ],
                "summary": "Загрузка нового документа",
                "operationId": "storeDocument",
                "requestBody": {
                    "description": "Данные для загрузки документа. Используйте multipart/form-data.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Название документа",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "Файл документа (PDF, DOCX, XLSX)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_for_all_employees": {
                                        "description": "Назначить всем сотрудникам",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Документ успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/documents/{document}": {
            "post": {
                "tags": [
                    "Администратор - Документы"
                ],
                "summary": "Обновление документа (включая замену файла)",
                "description": "Позволяет изменить название, флаг и/или заменить файл. Статусы ознакомления не сбрасываются. Используйте multipart/form-data и метод POST с полем _method=PUT.",
                "operationId": "updateDocument",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID документа для обновления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Обязательно для эмуляции PUT-запроса.",
                                        "type": "string",
                                        "default": "PUT",
                                        "enum": [
                                            "PUT"
                                        ]
                                    },
                                    "title": {
                                        "description": "Новое название (опционально)",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "Новый файл (опционально)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_for_all_employees": {
                                        "description": "Новое значение флага (опционально)",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Документ успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Документ не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Администратор - Документы"
                ],
                "summary": "Удаление документа",
                "operationId": "deleteDocument",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "description": "ID документа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Документ успешно удален"
                    },
                    "404": {
                        "description": "Документ не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/statistics": {
            "get": {
                "tags": [
                    "Администратор - Статистика"
                ],
                "summary": "Получение статистики по ознакомлению сотрудников",
                "operationId": "getStatistics",
                "responses": {
                    "200": {
                        "description": "Успешный ответ со статистикой",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserStatistics"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentStatus": {
                "title": "Document Status",
                "description": "Статус одного документа в списке статистики",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Правила компании"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "assigned",
                            "read"
                        ],
                        "example": "read"
                    }
                },
                "type": "object"
            },
            "UserStatistics": {
                "title": "User Statistics",
                "description": "Статистика по документам для одного сотрудника",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Батырев"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "К.В."
                    },
                    "patronymic": {
                        "type": "string",
                        "example": "",
                        "nullable": true
                    },
                    "position": {
                        "type": "string",
                        "example": "Бармен"
                    },
                    "avatar_url": {
                        "description": "Полный URL аватара",
                        "type": "string",
                        "nullable": true
                    },
                    "documents": {
                        "description": "Список назначенных документов и их статусы",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DocumentStatus"
                        }
                    }
                },
                "type": "object"
            },
            "Document": {
                "title": "Document",
                "description": "Модель документа",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "Название документа",
                        "type": "string",
                        "example": "Инструкция по технике безопасности"
                    },
                    "description": {
                        "description": "Краткое описание",
                        "type": "string",
                        "nullable": true
                    },
                    "file_path": {
                        "description": "Путь к файлу в хранилище",
                        "type": "string",
                        "example": "documents/xyz.pdf"
                    },
                    "original_filename": {
                        "description": "Исходное имя файла",
                        "type": "string",
                        "example": "instrukciya.pdf"
                    },
                    "file_mime_type": {
                        "description": "MIME-тип файла",
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "file_size": {
                        "description": "Размер файла в байтах",
                        "type": "integer",
                        "example": "102400"
                    },
                    "is_for_all_employees": {
                        "description": "Флаг 'Назначить всем'",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите Bearer токен для доступа к защищенным эндпоинтам",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}